refactor: split prompt formatters
