+++ hooks/usePlayerActions.ts
@@ -30,7 +30,7 @@ import {
   buildItemChangeRecords,
   applyAllItemChanges,
 } from '../utils/gameLogicUtils';
-import { formatMainGameTurnPrompt } from '../utils/promptFormatters';
+import { formatMainGameTurnPrompt } from '../utils/promptFormatters/dialogue';
 import { structuredCloneGameState } from '../utils/cloneUtils';
 import { handleMapUpdates } from '../utils/mapUpdateHandlers';
