+++ services/corrections/character.ts
@@ -4,7 +4,7 @@
  */
 import { AdventureTheme, Character, MapNode } from '../../types';
 import { MAX_RETRIES } from '../../constants';
-import { formatKnownPlacesForPrompt } from '../../utils/promptFormatters';
+import { formatKnownPlacesForPrompt } from '../../utils/promptFormatters/map';
 import { callCorrectionAI, callMinimalCorrectionAI } from './base';
 
 /** Structure returned when correcting character details. */
